backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0
pydantic-settings>=2.0
sqlalchemy>=2.0
rq>=1.16
redis>=5.0
openai>=1.0
gitpython>=3.1
google-genai>=1.0
PyGithub>=2.0
requests>=2.31
